requirements.txt
Flask
psycopg2-binary
SQLAlchemy
Werkzeug
DateTime
setuptools
PyJWT